Break up functionality for Cardano
